\name{iCenJMH-package}
\alias{iCenJMH-package}
\alias{iCenJMH}
\docType{package}
\title{
\packageTitle{iCenJMH}
}
\description{
\packageDescription{iCenJMH}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{iCenJMH}
\packageIndices{iCenJMH}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{iCenJMH}

Maintainer: \packageMaintainer{iCenJMH}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
