% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCenJMMLSM.R
\name{iCenJMMLSM}
\alias{iCenJMMLSM}
\title{Joint Modelling for Continuous outcomes}
\usage{
iCenJMMLSM(
  Ydata = NULL,
  Tdata = NULL,
  long.formula = NULL,
  surv.formula = NULL,
  variance.formula = NULL,
  random = NULL,
  timeVar = NULL,
  iCen.info = NULL,
  maxiter = 1000,
  epsilon = 1e-04,
  epsilonH0 = 1e-04,
  pStol = 1e-06,
  quadpoint = 10,
  print.para = FALSE,
  initial.para = TRUE
)
}
\arguments{
\item{Ydata}{a longitudinal data frame in long format.}

\item{Tdata}{a survival data frame.
Each subject has one data entry.}

\item{long.formula}{a formula object with the response variable and time-independent fixed effects covariates
to be included in the longitudinal sub-model.}

\item{surv.formula}{a formula object with the survival time, event indicator, and the time-independent fixed effects covariates
to be included in the survival sub-model.}

\item{variance.formula}{an one-sided formula object with the fixed effects covariates to model the variance of longituidnal sub-model.}

\item{random}{a one-sided formula object describing the random effects part of the longitudinal sub-model.
For example, fitting a random intercept model takes the form ~ 1|ID.
Alternatively. Fitting a random intercept and slope model takes the form ~ x1 + ... + xn|ID.}

\item{timeVar}{a string of an interval-censored time covariate.}

\item{maxiter}{the maximum number of iterations of the EM algorithm that the function will perform. Default is 10000.}

\item{epsilon}{Tolerance parameter for parametric components. Default is 0.0001.}

\item{epsilonH0}{Tolerance parameter for non-parametric Breslow estimator for the event of interest. Default is 0.0001.}

\item{pStol}{Tolerance parameter for the posterior probability of an initial event time Si. 
If the posterior probability is less than a specified value, then override the probability and move forward to 
the next probability in the E-step. Default is 1e-6.}

\item{quadpoint}{the number of Gauss-Hermite quadrature points. Default is 20.}

\item{print.para}{Print detailed information of each iteration. Default is FALSE, i.e., not to print the iteration details.}

\item{initial.para}{Input initial estimate of parameters. Default is FALSE.}

\item{c}{tuning parameter for bandwidth of kernel estimate of hazards.}
}
\description{
Joint modeling of longitudinal continuous data and competing risks
}
