% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmp_status.R
\name{openmp_status}
\alias{openmp_status}
\title{Check OpenMP status}
\usage{
openmp_status()
}
\value{
An integer.
}
\description{
Returns the number of OpenMP threads available at compile time.
A value greater than 1 usually indicates OpenMP is enabled.
A value of 1 indicates serial execution.
}
