% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.iCenjmcs.R
\name{print.iCenjmcs}
\alias{print.iCenjmcs}
\title{Print iCenjmcs}
\usage{
\method{print}{iCenjmcs}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object of class 'iCenjmcs'.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print contents of iCenjmcs object.
}
\seealso{
\code{\link{iCenjmcs}}
}
\author{
Shanpeng Li
}
