% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.iCenJMMLSM.R
\name{print.iCenJMMLSM}
\alias{print.iCenJMMLSM}
\title{Print iCenJMMLSM}
\usage{
\method{print}{iCenJMMLSM}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object of class 'iCenJMMLSM'.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print contents of iCenJMMLSM object.
}
\seealso{
\code{\link{iCenJMMLSM}}
}
\author{
Shanpeng Li
}
