% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.survfitiCenJMMLSM.R
\name{print.survfitiCenJMMLSM}
\alias{print.survfitiCenJMMLSM}
\title{Print survfitiCenJMMLSM}
\usage{
\method{print}{survfitiCenJMMLSM}(x, ...)
}
\arguments{
\item{x}{x of class 'survfitiCenJMMLSM'.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
Print survfitiCenJMMLSM
}
\seealso{
\code{\link{iCenJMMLSM}, \link{survfitiCenJMMLSM}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
