% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.BrieriCenJMMLSM.R
\name{summary.BrieriCenJMMLSM}
\alias{summary.BrieriCenJMMLSM}
\title{Print BrieriCenJMMLSM}
\usage{
\method{summary}{BrieriCenJMMLSM}(object, digits = 4, ...)
}
\arguments{
\item{object}{object of class 'BrieriCenJMMLSM'.}

\item{digits}{number of digits of decimal to be printed.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
Print BrieriCenJMMLSM
}
\seealso{
\code{\link{JMMLSM}, \link{survfitJMMLSM}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
