% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.DynPredAcciCenJMMLSM.R
\name{summary.DynPredAcciCenJMMLSM}
\alias{summary.DynPredAcciCenJMMLSM}
\title{Print DynPredAcciCenJMMLSM}
\usage{
\method{summary}{DynPredAcciCenJMMLSM}(object, digits = 4, ...)
}
\arguments{
\item{object}{object of class 'DynPredAcciCenJMMLSM'.}

\item{digits}{number of digits of decimal to be printed.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
Print DynPredAcciCenJMMLSM
}
\seealso{
\code{\link{iCenJMMLSM}, \link{survfitiCenJMMLSM}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
