% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.DynPredJMMLSM.R
\name{summary.DynPredJMMLSM}
\alias{summary.DynPredJMMLSM}
\title{Print DynPredJMMLSM}
\usage{
\method{summary}{DynPredJMMLSM}(object, digits = 4, ...)
}
\arguments{
\item{object}{object of class 'DynPredJMMLSM'.}

\item{digits}{number of digits of decimal to be printed.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
Print DynPredJMMLSM
}
\seealso{
\code{\link{iCenJMMLSM}, \link{survfitiCenJMMLSM}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
