% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.MAEQiCenJMMLSM.R
\name{summary.MAEQiCenJMMLSM}
\alias{summary.MAEQiCenJMMLSM}
\title{Print MAEQiCenJMMLSM}
\usage{
\method{summary}{MAEQiCenJMMLSM}(object, digits = 3, ...)
}
\arguments{
\item{object}{object of class 'MAEQiCenJMMLSM'.}

\item{digits}{number of decimal points to be rounded.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a list of matrices with conditional probabilities for subjects.
}
\description{
Print MAEQiCenJMMLSM
}
\seealso{
\code{\link{iCenJMMLSM}, \link{survfitiCenJMMLSM}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
