% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.DynPredAcciCenJMMLSM.R,
%   R/summary.iCenJMMLSM.R
\name{summary}
\alias{summary}
\alias{summary.DynPredAcciCenJMMLSM}
\alias{summary.iCenJMMLSM}
\title{Print evaluation metrics for the fitted joint model}
\usage{
\method{summary}{DynPredAcciCenJMMLSM}(object, digits = 4, ...)

\method{summary}{iCenJMMLSM}(object, process = c("longitudinal", "survival"), digits = 4, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{iCenJMMLSM}.}

\item{digits}{the number of significant digits to use when printing. Default is 4.}

\item{...}{further arguments passed to or from other methods.}

\item{process}{for which model (i.e., longitudinal model or survival model) to extract the estimated coefficients.}
}
\value{
a list of matrices with conditional probabilities for subjects.

A table to summarize the model results.
}
\description{
Produce result summaries of a joint model fit.
}
\seealso{
\code{\link{iCenJMMLSM}, \link{survfitiCenJMMLSM}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
